chore: migrated the README and BACKLOG, as well as the CLI.
CLI is not applicable for an Astro project and will have to be refactored accordingly.

I am currently working to push the updated version of the CLI with a better documentation on it as well.

The goal is to get a CLI that is inspirated by node ace cli, preferably with as many commands as possible with protections against destructive db commands